training pipeline partially working + notebook
