+++ airflow/dags/scripts/tasks.py
@@ -1,17 +1,26 @@
 import json
 import pandas as pd
+from sklearn.metrics import accuracy_score
+from sklearn.metrics import f1_score
 from sklearn.model_selection import train_test_split
-from typing import List
-from churn_library.processing_helpers import (
-    find_columns_for_imputation, 
-    impute_with_constant, 
+from churn_library.helpers import (
     create_null_flags,
     remove_rows_with_any_null,
-    clean_negatives)
-
-from churn_library.feature_helpers import (create_new_features, 
-                                           binarize_and_align_quantiles,
-                                           detect_and_convert_categoricals)
+    clean_negatives,
+    detect_and_convert_categoricals,
+    create_new_features,
+    sanitize_column_names)
+#import mlflow
+#import mlflow.sklearn
+import pickle
+from sklearn.model_selection import cross_val_score
+from sklearn.ensemble import RandomForestClassifier
+from sklearn.linear_model import LogisticRegression
+from sklearn.neighbors import KNeighborsClassifier
+from lightgbm import LGBMClassifier
+from xgboost import XGBClassifier
+import joblib
+from typing import Dict
 
 #======PROCESSING TASKS=========
 def remove_rows_with_any_null_task(input_path, output_path, config):
@@ -38,39 +47,9 @@ def clean_negatives_task(input_path, output_path, config):
     exception_cols = config['preprocessing']['negative_values']['exception_cols']
 
     df_processed = clean_negatives(df, exception_cols=exception_cols)
-    
+
     df_processed.to_csv(output_path, sep=';', index=False)
     
-    print(f"Task finished. Data saved to {output_path}")
-
-
-def impute_zeros_task(input_path, output_path, config):
-    """Airflow task: reads CSV, imputes with 0 on specified columns, and writes CSV."""
-    print("--- Starting zero imputation task ---")
-    df = pd.read_csv(input_path, sep=';')
-
-    # --- LÍNEAS DE DEPURACIÓN ---
-    print("\n--- DEBUGGING: Column Search ---")
-    print(f"Total columns available in DataFrame ({len(df.columns)}):")
-    print(df.columns.tolist())
-    
-    imputation_config = config.get('preprocessing', {}).get('imputation', {})
-    
-    zero_fill_config = {
-        'patterns': imputation_config.get('zero_fill_patterns', {}),
-        'exact_columns': imputation_config.get('zero_fill_exact_columns', [])
-    }
-    print(f"\nConfiguration being used for search: {zero_fill_config}")
-    print("--- END DEBUGGING ---\n")
-    # ------------------------------------
-
-    # Usamos la función inteligente de la librería para encontrar las columnas
-    cols_for_zero = find_columns_for_imputation(df, zero_fill_config)
-
-    print(f"Found {len(cols_for_zero)} columns to impute with zero: {cols_for_zero}")
-
-    df_processed = impute_with_constant(df, cols_for_zero, fill_value=0)
-    df_processed.to_csv(output_path, sep=';', index=False)
     print(f"Task finished. Data saved to {output_path}")
 
 
@@ -101,38 +80,6 @@ def calculate_imputation_values_task(input_path, output_path_json, config):
         
     print("Task finished.")
 
-
-def apply_imputation_task(input_path_data, output_path_data, input_path_json):
-    """
-    Airflow task: reads a dataset and the saved imputation values from a JSON file,
-    then applies the imputation.
-    This task "transforms" our data.
-    """
-    print(f"--- Applying imputation from {input_path_json} to {input_path_data} ---")
-    df = pd.read_csv(input_path_data, sep=';')
-    
-    print("Loading imputation values...")
-    with open(input_path_json, 'r') as f:
-        imputation_values = json.load(f)
-        
-    print(f"Applying median imputation for columns: {list(imputation_values.keys())}")
-    # Usamos .fillna directamente con el diccionario, es más eficiente
-    df.fillna(value=imputation_values, inplace=True)
-            
-    df.to_csv(output_path_data, sep=';', index=False)
-    print(f"Task finished. Transformed data saved to {output_path_data}")
-
-
-def impute_categoricals_task(input_path, output_path, config):
-    """Airflow task: reads CSV, imputes categorical columns with 'UNKN', and writes CSV."""
-    print("--- Starting categorical imputation task ---")
-    df = pd.read_csv(input_path, sep=';')
-    fill_value = config['preprocessing']['imputation']['categorical_fill_value']
-    cols_for_unkn = df.select_dtypes(include=['object', 'category']).columns.tolist()
-    # We use your helper function to impute
-    df_processed = impute_with_constant(df, cols_for_unkn, fill_value=fill_value)
-    df_processed.to_csv(output_path, sep=';', index=False)
-    print(f"Task finished. Data saved to {output_path}")
 
 
 def create_null_flags_task(input_path, output_path, config):
@@ -165,27 +112,16 @@ def train_test_split_task(input_path, train_output_path, test_output_path, confi
         X, y, test_size=config['preprocessing']['imputation']['test_size'], random_state=config['random_state']
     )
 
-    train_df = pd.concat([X_train, y_train], axis=1)
-    test_df = pd.concat([X_test, y_test], axis=1)
+    df_train = pd.concat([X_train, y_train], axis=1)
+    df_test = pd.concat([X_test, y_test], axis=1)
 
     print(f"Guardando set de entrenamiento en: {train_output_path}")
-    train_df.to_csv(train_output_path, sep=';', index=False)
+    df_train.to_csv(train_output_path, sep=';', index=False)
     
     print(f"Guardando set de prueba en: {test_output_path}")
-    test_df.to_csv(test_output_path, sep=';', index=False)
+    df_test.to_csv(test_output_path, sep=';', index=False)
 
     print("División de datos completada.")
-
-def impute_with_unkn_task(input_path, output_path, config):
-    """Airflow task: reads CSV, imputes categorical columns with 'UNKN', and writes CSV."""
-    print("--- Starting categorical imputation task ---")
-    df = pd.read_csv(input_path, sep=';')
-    fill_value = config['preprocessing']['imputation']['categorical_fill_value']
-    cols_for_unkn = config['preprocessing']['imputation']['unkn_fill_exact_columns']
-    # We use your helper function to impute
-    df_processed = impute_with_constant(df, cols_for_unkn, fill_value=fill_value)
-    df_processed.to_csv(output_path, sep=';', index=False)
-    print(f"Task finished. Data saved to {output_path}")
 
 
 def remove_columns_task(input_path, output_path, config):
@@ -223,13 +159,24 @@ def impute_all_nulls_task(input_path, output_path, imputation_values_path, confi
 
     # 1. Imputar columnas categóricas con el valor constante 'UNKN'
     fill_value = config['preprocessing']['imputation']['categorical_fill_value']
-    cols_for_unkn = config['preprocessing']['imputation']['unkn_fill_exact_columns']
     
-    for col in cols_for_unkn:
-        df[col].fillna(fill_value, inplace=True)
-        
-    print(f"Imputación de columnas categóricas completada con el valor: '{fill_value}'")
+    df = detect_and_convert_categoricals(df)
 
+    # Identificar todas las columnas categóricas (object y category) en el DataFrame
+    categorical_cols = df.select_dtypes(include=['object', 'category']).columns.tolist()
+    
+    print(f"Columnas categóricas a imputar: {categorical_cols}")
+
+    for col in categorical_cols:
+        # --- Solución clave: Añadir la nueva categoría antes de rellenar ---
+        # Si la columna es de tipo 'category', añadir la nueva categoría 'UNKN'
+        if pd.api.types.is_categorical_dtype(df[col]):
+            df[col] = df[col].cat.add_categories([fill_value])
+        
+        # Ahora, rellenar los nulos
+        df[col].fillna(fill_value, inplace=True)
+
+    print(f"Imputación de TODAS las columnas categóricas completada con el valor: '{fill_value}'")
     # 2. Imputar columnas numéricas con la mediana
     columns_impute_median = config['preprocessing']['imputation']['median_fill_columns']
 
@@ -242,9 +189,14 @@ def impute_all_nulls_task(input_path, output_path, imputation_values_path, confi
     
     # Verificar que no queden nulos después de la imputación
     print("\nVerificación final de nulos:")
-    print(df[cols_for_unkn + columns_impute_median].isnull().sum())
+    print(df[categorical_cols + columns_impute_median].isnull().sum())
     
     # Guardar el DataFrame final
+    nulos_por_columna = df.isnull().sum()
+
+    # Filtramos para mostrar solo las columnas que tienen al menos un valor nulo
+    columnas_con_nulos = nulos_por_columna[nulos_por_columna > 0]
+    print(f"NUMERO DE NULOS DESPUES DE PROCESAR {columnas_con_nulos}")
     df.to_csv(output_path, sep=';', index=False)
     print(f"\nTarea finalizada. Datos guardados en: {output_path}")
 
@@ -322,99 +274,309 @@ def create_new_features_task(input_path_train, input_path_test, output_path_trai
     print("Feature engineering task completed successfully.")
 
 
-def one_hot_encode_and_align_task(input_path_train, input_path_test, output_path_train, output_path_test, config):
+def binarize_train_set_task(input_path, output_path, columns, q: int, bins_path: str) -> pd.DataFrame:
     """
-    Reads the train and test sets, applies One-Hot Encoding, and aligns their columns.
+    Binariza las columnas del set de entrenamiento y guarda los límites de los bins.
     """
-    print("Starting the One-Hot Encoding and column alignment task...")
+    print("Starting training set binning...")
     
-    # Read the DataFrames
-    df_train = pd.read_csv(input_path_train, sep=';')
-    df_test = pd.read_csv(input_path_test, sep=';')
+    # Creamos una copia para evitar SettingWithCopyWarning
+    df_train = pd.read_csv(input_path, sep=';')
+    df_binned = df_train.copy()
+    bins_map = {}
     
-    # Get the list of categorical variables to encode
-    categorical_cols_to_encode = config['preprocessing']['categorical_for_ohe'] # Assumed this list is in the config
+    for col in columns:
+        # Usamos pd.qcut para obtener los bins y los límites
+        df_binned[f'{col}_binned'], bins = pd.qcut(
+            df_binned[col],
+            q=q,
+            labels=False, # Usamos etiquetas numéricas
+            retbins=True,
+            duplicates='drop'
+        )
+        bins_map[col] = bins
+        
+        # Elimina la columna original
+        df_binned.drop(columns=[col], inplace=True)
 
-    # Perform One-Hot Encoding on the train set
-    df_train_encoded = pd.get_dummies(df_train, columns=categorical_cols_to_encode, drop_first=True)
+    # Guardar el mapeo de los bins para usarlo en el set de prueba
+    print(f"Saving bins in: {bins_path}")
+    joblib.dump(bins_map, bins_path)
     
-    # Save the list of ALL resulting columns
-    final_train_cols = df_train_encoded.columns.tolist()
+    print("Finished training set binning.")
+    df_binned.to_csv(output_path, sep=';', index=False)
 
-    # Perform One-Hot Encoding on the test set
-    df_test_encoded = pd.get_dummies(df_test, columns=categorical_cols_to_encode, drop_first=True)
-    
-    # Align the columns of the test set with the train set
-    df_test_aligned = df_test_encoded.reindex(columns=final_train_cols, fill_value=0)
-    
-    print(f"Number of columns in the train set: {df_train_encoded.shape[1]}")
-    print(f"Number of columns in the aligned test set: {df_test_aligned.shape[1]}")
+    return df_binned
 
-    # Save the processed DataFrames
-    df_train_encoded.to_csv(output_path_train, sep=';', index=False)
-    df_test_aligned.to_csv(output_path_test, sep=';', index=False)
-
-    print("One-Hot Encoding and alignment task completed successfully.")
-
-
-def binarize_and_align_quantiles_task(input_path_train: str, input_path_test: str, output_path_train: str, output_path_test: str, columns_to_bin: List[str], q: int):
+# Nueva función para aplicar el binarizador en el set de prueba
+def binarize_test_set_task(input_path, output_path, columns, bins_path: str) -> pd.DataFrame:
     """
-    Lee los sets de entrenamiento y prueba, binariza y alinea sus columnas.
+    Aplica los límites de los bins guardados en el set de prueba.
     """
-    print(f"Leyendo set de entrenamiento desde: {input_path_train}")
-    df_train = pd.read_csv(input_path_train, sep=';')
+
+    print(f"Reading data from {input_path} and binarizing columns: {columns}")
+
+    # Step 1: Read the data from the input path
+    df_test = pd.read_csv(input_path, sep=';')
+
+    print("Starting test set binning...")
     
-    print(f"Leyendo set de prueba desde: {input_path_test}")
-    df_test = pd.read_csv(input_path_test, sep=';')
+    # Cargamos el mapeo de bins guardado
+    print(f"Loading bins...: {bins_path}")
+    bins_map = joblib.load(bins_path)
+
+    df_binned = df_test.copy()
     
-    # Llama a tu función principal para binarizar y alinear
-    df_train_binned, df_test_binned = binarize_and_align_quantiles(df_train, df_test, columns_to_bin, q)
-    
-    # Guardar los DataFrames procesados
-    df_train_binned.to_csv(output_path_train, sep=';', index=False)
-    df_test_binned.to_csv(output_path_test, sep=';', index=False)
-    
-    print("Tarea de binarización y alineación completada con éxito.")
+    for col in columns:
+        bins = bins_map[col]
+        df_binned[f'{col}_binned'] = pd.cut(
+            df_binned[col],
+            bins=bins,
+            labels=False, 
+            right=True,
+            include_lowest=True
+        )
+
+        df_binned.drop(columns=[col], inplace=True)
+        
+    print("Test set binning completed.")
+    df_binned.to_csv(output_path, sep=';', index=False)
+
+    return df_binned
 
 def one_hot_encode_and_align_task(input_path_train, input_path_test, output_path_train, output_path_test):
     """
-    Reads the train and test sets, automatically detects categorical variables,
-    applies One-Hot Encoding, and aligns their columns.
+    Lee los sets de entrenamiento y prueba, detecta variables categóricas,
+    aplica One-Hot Encoding y alinea sus columnas.
     """
     print("Starting the One-Hot Encoding and column alignment task...")
     
-    # Read the DataFrames
+    # 1. Lee los DataFrames
     df_train = pd.read_csv(input_path_train, sep=';')
     df_test = pd.read_csv(input_path_test, sep=';')
     
-    # Detect and convert categorical columns on the train set
-    df_train_converted, converted_cols = detect_and_convert_categoricals(df_train)
+    # 2. Sanitiza los nombres de las columnas
+    # Esto es crucial para evitar errores en pasos posteriores
+    df_train_clean = sanitize_column_names(df_train)
+    df_test_clean = sanitize_column_names(df_test)
     
-    # Apply the same conversion to the test set using the list from the train set
-    df_test_converted, _ = detect_and_convert_categoricals(df_test)
-    
-    # Get the list of ALL categorical variables (both object and converted)
+    # 3. Detecta y convierte columnas numéricas discretas a categóricas
+    # Esta llamada crea las variables 'df_train_converted' y 'df_test_converted'
+    df_train_converted = detect_and_convert_categoricals(df_train_clean)
+    df_test_converted = detect_and_convert_categoricals(df_test_clean)
+
+    # 4. Obtiene la lista de TODAS las columnas categóricas para codificar
     categorical_cols_to_encode = df_train_converted.select_dtypes(include=['object', 'category']).columns.tolist()
 
     print(f"Applying One-Hot Encoding to the following columns: {categorical_cols_to_encode}")
 
-    # Perform One-Hot Encoding on the train set
+    # 5. Realiza One-Hot Encoding en el set de entrenamiento
     df_train_encoded = pd.get_dummies(df_train_converted, columns=categorical_cols_to_encode, drop_first=True)
     
-    # Perform One-Hot Encoding on the test set
+    # 6. Realiza One-Hot Encoding en el set de prueba
     df_test_encoded = pd.get_dummies(df_test_converted, columns=categorical_cols_to_encode, drop_first=True)
     
-    # Save the list of ALL resulting columns from the training set
+    # 7. Alinea las columnas del set de prueba con el de entrenamiento
     final_train_cols = df_train_encoded.columns.tolist()
-
-    # Align the columns of the test set with the train set, filling missing columns with 0
     df_test_aligned = df_test_encoded.reindex(columns=final_train_cols, fill_value=0)
     
     print(f"Number of columns in the train set: {df_train_encoded.shape[1]}")
     print(f"Number of columns in the aligned test set: {df_test_aligned.shape[1]}")
 
-    # Save the processed DataFrames
+    # 8. Guarda los DataFrames procesados
     df_train_encoded.to_csv(output_path_train, sep=';', index=False)
     df_test_aligned.to_csv(output_path_test, sep=';', index=False)
-
+    
     print("One-Hot Encoding and alignment task completed successfully.")
+
+    return {
+        'df_train_path': output_path_train,
+        'df_test_path': output_path_test
+    }
+
+
+#===========TRAINING TASK
+
+def train_model_task(input_path: str, model_config: dict, target_variable_name: str):#, columns_to_drop):
+    """
+    Tarea que entrena un modelo de Machine Learning.
+    """
+    print("Starting the model training task...")
+    
+    df = pd.read_csv(input_path, sep=';')
+
+    #df.drop(columns=columns_to_drop, errors='ignore', inplace=True)
+
+    X = df.drop(columns=[target_variable_name])
+    y = df[target_variable_name]
+    
+    # Use the passed-in model_config directly
+    model_name = model_config.get('model_name')
+    model_params = model_config.get('params', {})
+    cv_folds = model_config.get('cv_folds')
+    
+    final_model = None
+
+    if model_name == 'RandomForestClassifier':
+        model = RandomForestClassifier(random_state=42, **model_params)
+    elif model_name == 'LogisticRegression':
+        model = LogisticRegression(random_state=42, **model_params)
+    elif model_name == 'KNeighborsClassifier':
+        model = KNeighborsClassifier(**model_params)
+    elif model_name == 'LGBMClassifier':
+        model = LGBMClassifier(random_state=42, **model_params)
+    elif model_name == 'XGBClassifier':
+        model = XGBClassifier(random_state=42, use_label_encoder=False, eval_metric='logloss', **model_params)
+    else:
+        raise ValueError(f"Unsupported model: {model_name}")
+
+    if cv_folds and cv_folds > 1:
+        print(f"Starting cross-validation with {cv_folds} folds for {model_name}...")
+        scores = cross_val_score(model, X, y, cv=cv_folds, scoring='accuracy')
+        avg_accuracy = scores.mean()
+        print(f"\nCross-validation completed. Average Accuracy: {avg_accuracy:.4f}")
+        print(f"Retraining the {model_name} on the full training dataset...")
+        final_model = model
+        final_model.fit(X, y)
+        
+    else:
+        print(f"No cross-validation specified. Training {model_name} on the full dataset...")
+        final_model = model
+        final_model.fit(X, y)
+
+    model_name = model_config['model_name']
+    model_file_path = f"/opt/airflow/models/{model_name}_trained.pkl"
+
+    # Guarda el modelo en un archivo usando joblib
+    print(f"Guardando el modelo en: {model_file_path}")
+    joblib.dump(final_model, model_file_path)
+
+    # Devuelve la ruta del archivo, no el objeto del modelo
+    return model_file_path
+    
+
+def evaluate_model_task(model_path: str, test_data_path: str, target_variable: str):
+    """
+    Tarea que lee un modelo y datos de prueba para evaluar su rendimiento.
+    """
+    print(f"Starting model evaluation from {test_data_path}...")
+    
+    try:
+        with open(model_path, 'rb') as f:
+            model = pickle.load(f)
+        
+        df_test = pd.read_csv(test_data_path, sep=';')
+        
+    except FileNotFoundError:
+        print("Error: Files not found. Ensure the training pipeline ran successfully.")
+        return
+
+    X_test = df_test.drop(columns=[target_variable])
+    y_test = df_test[target_variable]
+    
+    y_pred = model.predict(X_test)
+    accuracy = accuracy_score(y_test, y_pred)
+    f1 = f1_score(y_test, y_pred)
+    
+    print("\n--- Model Evaluation Results ---")
+    print(f"Accuracy: {accuracy:.4f}")
+    print(f"F1-Score: {f1:.4f}")
+    print("--------------------------------")
+
+def evaluate_model_task(model_path: str, test_data_path: str, target_variable: str):
+    """
+    Tarea que lee un modelo entrenado y datos de prueba para evaluar su rendimiento.
+    
+    Args:
+        model_path (str): Ruta al modelo entrenado en formato .pkl.
+        test_data_path (str): Ruta al archivo de datos de prueba.
+        target_variable (str): Nombre de la columna de la variable objetivo.
+    """
+    print(f"Starting model evaluation from {test_data_path}...")
+    
+    # 1. Cargar el modelo del archivo
+    try:
+        model = joblib.load(model_path)
+        print("Model loaded successfully.")
+    except FileNotFoundError:
+        print(f"Error: Model file not found at {model_path}. "
+              "Ensure the training task ran successfully.")
+        return # Salir de la función si el archivo no se encuentra
+
+    # 2. Cargar los datos de prueba
+    try:
+        df_test = pd.read_csv(test_data_path, sep=';')
+        print("Test data loaded successfully.")
+    except FileNotFoundError:
+        print(f"Error: Test data file not found at {test_data_path}.")
+        return
+
+    # 3. Separar las características (X) y la variable objetivo (y)
+    X_test = df_test.drop(columns=[target_variable])
+    y_test = df_test[target_variable]
+    
+    # 4. Generar predicciones y calcular métricas
+    y_pred = model.predict(X_test)
+    accuracy = accuracy_score(y_test, y_pred)
+    f1 = f1_score(y_test, y_pred)
+    
+    # 5. Imprimir los resultados de la evaluación
+    print("\n--- Model Evaluation Results ---")
+    print(f"Accuracy: {accuracy:.4f}")
+    print(f"F1-Score: {f1:.4f}")
+    print("--------------------------------")
+
+def predict_task(model_path: str, new_data_path: str, output_path: str, config):
+    """
+    Tarea que lee un modelo entrenado y genera predicciones sobre nuevos datos.
+
+    Args:
+        model_path (str): Ruta al modelo entrenado en formato .pkl.
+        new_data_path (str): Ruta al archivo de datos nuevos a predecir.
+        output_path (str): Ruta donde se guardarán las predicciones.
+    """
+    print(f"Starting prediction task...")
+
+    try:
+        # Lee el modelo ya entrenado
+        with open(model_path, 'rb') as f:
+            model = pickle.load(f)
+
+        # Lee los datos nuevos (asume que están limpios y sanitizados)
+        new_data_df = pd.read_csv(new_data_path, sep=';')
+        
+    except FileNotFoundError:
+        print("Error: Files not found. Ensure model and data paths are correct.")
+        return
+    
+    target = config['general']['target']
+    if target in new_data_df.columns:
+        print("Target variable found. Performing model evaluation...")
+        X = new_data_df.drop(columns=[target])
+        y_true = new_data_df[target]
+        
+        predictions = model.predict(X)
+        
+        accuracy = accuracy_score(y_true, predictions)
+        f1 = f1_score(y_true, predictions)
+        
+        print("\n--- Model Evaluation Results ---")
+        print(f"Accuracy: {accuracy:.4f}")
+        print(f"F1-Score: {f1:.4f}")
+        print("--------------------------------")
+        
+        predictions_df = pd.DataFrame(new_data_df.copy())
+        predictions_df['prediction'] = predictions
+        
+    else:
+        print("Target variable not found. Generating predictions without evaluation.")
+        predictions = model.predict(new_data_df)
+        predictions_df = pd.DataFrame(new_data_df.copy())
+        predictions_df['prediction'] = predictions
+
+    # Guarda el DataFrame con las predicciones
+    try:
+        predictions_df.to_csv(output_path, sep=';', index=False)
+        print(f"Predictions saved successfully to {output_path}")
+    except Exception as e:
+        print(f"Error saving predictions to {output_path}: {e}")
