--- churn_library/feature_helpers.py
@@ -1,160 +1,0 @@
-import numpy as np
-import pandas as pd
-from pandas.api.types import is_numeric_dtype
-from typing import Optional, List, Tuple
-
-def group_category_by_churn(df: pd.DataFrame, column: str, upper_th: float, lower_th: float, new_column_name: str, target: str = 'churn'):
-    df_copia = df.copy()
-
-    churn_by_category = df_copia.groupby(column)[target].mean().reset_index()
-    churn_by_category = churn_by_category.rename(columns={target: 'mean_churn'})
-
-    # 2. Asignar la categoría de riesgo
-    def asignar_riesgo(tasa):
-        if tasa >= upper_th:
-            return 'Alto Riesgo'
-        elif tasa <= lower_th:
-            return 'Bajo Riesgo'
-        else:
-            return 'Riesgo Moderado'
-
-    churn_by_category[new_column_name] = churn_by_category['tasa_churn_media'].apply(asignar_riesgo)
-    
-    # 3. Crear el mapeo de categoría a grupo de riesgo
-    mapeo = churn_by_category.set_index(column)[new_column_name]
-
-    # 4. Unir la nueva columna de riesgo al DataFrame original
-    df_final = pd.merge(df_copia, churn_by_category[[column, new_column_name]], on=column, how='left')
-
-    return df_final, mapeo
-
-def detect_and_convert_categoricals(df: pd.DataFrame, umbral_discreta: int = 20):
-    """
-    Detects numeric columns that are actually discrete/categorical
-    and converts them to the 'category' type.
-    
-    Parameters:
-    df (pd.DataFrame): The DataFrame containing the data.
-    umbral_discreta (int): The number of unique values below which
-                           a numeric variable is considered categorical.
-    
-    Returns:
-    tuple: A tuple with (the DataFrame with conversions, the list of converted columns).
-    """
-    df_copy = df.copy()
-    variables_convertidas = []
-
-    for col in df_copy.columns:
-        if is_numeric_dtype(df_copy[col]) and col != 'churn':
-            num_valores_unicos = df_copy[col].nunique()
-
-            if num_valores_unicos <= umbral_discreta:
-                df_copy[col] = df_copy[col].astype('category')
-                variables_convertidas.append(col)
-    
-    print(f"Conversion completed. Converted {len(variables_convertidas)} columns to 'category' type.")
-    print("Converted columns:", variables_convertidas)
-    
-    return df_copy, variables_convertidas
-
-
-def get_quantile_bins(df: pd.DataFrame, column: str, q: int = 5, labels: Optional[List[str]] = None) -> Tuple[pd.DataFrame, Optional[List]]:
-    """
-    Bins a numeric column into quantiles, creating a new categorical column
-     and returns the list of bins for use on a different dataset.
-    
-    Args:
-    df (pd.DataFrame): The DataFrame containing the data.
-    column (str): The name of the numeric column to bin.
-    q (int): The number of quantiles (bins) for the discretization.
-    labels (List[str], optional): A list of labels for the bins.
-                                    If None, default labels will be used.
-
-    Returns:
-    Tuple: A tuple containing:
-           - pd.DataFrame: The DataFrame with the new binned column.
-           - Optional[List]: The list of bins generated by pd.qcut.
-    """
-    if column not in df.columns:
-        print(f"Error: Column '{column}' not found in the DataFrame.")
-        return df, None
-    if not is_numeric_dtype(df[column]):
-        print(f"Error: Column '{column}' is not numeric.")
-        return df, None
-
-    df_copy = df.copy()
-    binned_column_name = f'{column}_binned'
-
-    if labels and len(labels) != q:
-        print(f"Warning: The length of labels ({len(labels)}) does not match the number of quantiles ({q}). Default labels will be used.")
-        labels = None
-    
-    try:
-        # qcut returns the bins as a Categorical index
-        binned_series, bins = pd.qcut(df_copy[column], q=q, labels=labels, duplicates='drop', retbins=True)
-        df_copy[binned_series] = binned_series
-        
-        # Drop the original column to avoid multicollinearity
-        df_copy.drop(columns=[column], inplace=True)
-        
-        print(f"Column '{column}' was binned into {q} ranges and saved in '{binned_column_name}'.")
-        return df_copy, bins.tolist()
-    
-    except ValueError as e:
-        print(f"Warning: Could not bin column '{column}': {e}")
-        print(f"Column '{column}' will be kept as is.")
-        return df, None
-
-def binarize_and_align_quantiles(df_train: pd.DataFrame, df_test: pd.DataFrame, columns: List[str], q: int = 5) -> Tuple[pd.DataFrame, pd.DataFrame]:
-    """
-    Bins a set of columns in the train and test sets, using the quantiles
-    from the train set to align the transformation on the test set.
-    
-    Args:
-        df_train (pd.DataFrame): The training DataFrame.
-        df_test (pd.DataFrame): The testing DataFrame.
-        columns (List[str]): A list of numeric column names to bin.
-        q (int): The number of quantiles for the discretization.
-        
-    Returns:
-        Tuple: A tuple with (df_train_binned, df_test_binned).
-    """
-    df_train_binned = df_train.copy()
-    df_test_binned = df_test.copy()
-    
-    for col in columns:
-        print(f"--- Processing column '{col}' ---")
-        df_train_binned, bins_list = get_quantile_bins(df_train_binned, col, q=q)
-        
-        labels = [f'{col}_Q{i+1}' for i in range(q)] # Create default labels
-        
-        df_test_binned[f'{col}_binned'] = pd.cut(df_test_binned[col], bins=bins_list, labels=labels, right=True, include_lowest=True)
-        
-        df_test_binned.drop(columns=[col], inplace=True)
-        
-    print("\nBinning and alignment process completed.")
-    return df_train_binned, df_test_binned
-
-def create_new_features(df):
-    df_copy = df.copy()
-
-    # Overage
-    df_copy['total_overage_rev'] = df_copy['ovrrev_Mean'] + df_copy['vceovr_Mean'] + df_copy['datovr_Mean']
-    df_copy['overage_ratio'] = df_copy['total_overage_rev'] / df_copy['totmrc_Mean']
-
-    # Minute of use
-    df_copy['total_mou_Mean'] = df_copy['mou_cvce_Mean'] + df_copy['mou_cdat_Mean']
-    df_copy['complete_to_attempt_ratio'] = df_copy['complete_Mean'] / df_copy['attempt_Mean']
-
-    # Uso y pago
-    df_copy['mou_per_euro'] = df_copy['mou_Mean'] / df_copy['totmrc_Mean']
-    df_copy['relative_change_rev_mou'] = df_copy['change_rev'] / df_copy['change_mou']
-
-    variables_originales = [
-        'ovrrev_Mean', 'vceovr_Mean', 'datovr_Mean', 'totmrc_Mean',
-        'mou_cvce_Mean', 'mou_cdat_Mean', 'complete_Mean', 'attempt_Mean',
-        'mou_Mean', 'change_rev', 'change_mou'
-    ]
-    df_final = df_copy.drop(columns=variables_originales, errors='ignore')
-    
-    return df_final
