+++ docker-compose.yaml
@@ -81,7 +81,7 @@ x-airflow-common:
     - ${AIRFLOW_PROJ_DIR:-.}/data:/opt/airflow/data       
     - ${AIRFLOW_PROJ_DIR:-.}/src:/opt/airflow/src         
     - ${AIRFLOW_PROJ_DIR:-.}/models:/opt/airflow/models 
-    #- ./churn_library:/opt/airflow/churn_library
+    - ./churn_library:/opt/airflow/churn_library
   user: "${AIRFLOW_UID:-50000}:0"
   depends_on:
     &airflow-common-depends-on
