+++ requirements.txt
@@ -1,43 +1,5 @@
-asttokens==3.0.0
-comm==0.2.3
-contourpy==1.3.3
-cycler==0.12.1
-debugpy==1.8.16
-decorator==5.2.1
-executing==2.2.0
-fonttools==4.59.1
-ipykernel==6.30.1
-ipython==9.4.0
-ipython_pygments_lexers==1.1.1
-jedi==0.19.2
-jupyter_client==8.6.3
-jupyter_core==5.8.1
-kiwisolver==1.4.9
-matplotlib==3.10.5
-matplotlib-inline==0.1.7
-missingno==0.5.2
-nest-asyncio==1.6.0
-numpy==2.3.2
-packaging==25.0
-pandas==2.3.1
-parso==0.8.4
-pexpect==4.9.0
-pillow==11.3.0
-platformdirs==4.3.8
-prompt_toolkit==3.0.51
-psutil==7.0.0
-ptyprocess==0.7.0
-pure_eval==0.2.3
-Pygments==2.19.2
-pyparsing==3.2.3
-python-dateutil==2.9.0.post0
-pytz==2025.2
-pyzmq==27.0.1
-scipy==1.16.1
-seaborn==0.13.2
-six==1.17.0
-stack-data==0.6.3
-tornado==6.5.2
-traitlets==5.14.3
-tzdata==2025.2
-wcwidth==0.2.13
+pandas
+scikit-learn
+missingno
+xgboost
+lightgbm
